requirements.txt
python-telegram-bot>=21.0
google-genai>=1.66.0
Pillow>=10.0
python-dotenv>=1.0
